+++ apps/src/lib/utils/analytic.ts
@@ -1,13 +1,85 @@
 import type { ServerLoadEvent } from '@sveltejs/kit';
+// import { PostHog } from 'posthog-node';
+import posthog from 'posthog-js';
+import { browser } from '$app/environment';
 import { PostHog } from 'posthog-node';
+import { logger } from './logger';
 
-export const captureEvent = async (event: ServerLoadEvent, event_name: any) => {
-    const session = await event.locals.auth();
+/**
+ * Analytics utility functions for PostHog integration, Client setup. Call these functions in your server actions or routes.
+ * 
+ * This module provides functions to:
+ * - Capture events and send to PostHog
+ * - Identify users and set their properties
+ * 
+ * The analytics flow:
+ * 1. Events are captured with user ID and event name
+ * 2. Users are identified with properties like email, role, etc.
+ * 3. PostHog client is properly shutdown after each operation
+ */
 
-    const posthog = new PostHog(import.meta.env.VITE_POSTHOG_KEY, { host: 'https://us.i.posthog.com' });
-    const user = session?.user?.id ?? 'anonymous';
+export function initPostHog() {
+    if (browser) {
+        posthog.init(
+            import.meta.env.VITE_POSTHOG_KEY,
+            {
+                api_host: import.meta.env.VITE_POSTHOG_HOST,
+                person_profiles: 'always',
+                capture_pageview: true, // prevent double count pageviews and pageleaves
+                capture_pageleave: true, // prevent double count pageviews and pageleaves
+                persistence: 'localStorage'
+            }
+        );
+    }
+}
 
-    posthog.capture({ distinctId: user, event: event_name });
+/**
+ * Captures an analytics event and sends it to PostHog Server setup
+ * @param event - The SvelteKit server load event containing session info
+ * @param identityId - Optional unique identifier for the user
+ * @param eventName - Name of the event to capture
+ * @param properties - Additional properties for the event
+ */
+export const captureEventServer = async (identityId: string, eventName: string, properties?: Record<string, any>) => {
+
+    const posthog = new PostHog(import.meta.env.VITE_POSTHOG_KEY, { host: import.meta.env.VITE_POSTHOG_HOST });
+    posthog.capture({
+        distinctId: identityId ?? 'visitor',
+        event: eventName ?? 'page_view',
+        properties
+    });
 
     await posthog.shutdown();
 }
+
+export const captureEventClient = async (identityId: string, eventName: string, properties?: Record<string, any>) => {
+
+    if (!browser) {
+        return;
+    }
+
+    initPostHog();
+    posthog.identify(identityId ?? 'visitor');
+
+    posthog.capture(
+        eventName ?? 'page_view',
+        properties
+    );
+}
+
+
+/**
+ * Identifies a user in PostHog and sets their properties
+ * @param userId - Unique identifier for the user
+ * @param properties - Object containing user properties like email, role, etc.
+ */
+export const identifyUser = async (userId: string, properties: Record<string, any>) => {
+    const posthog = new PostHog(import.meta.env.VITE_POSTHOG_KEY, { host: import.meta.env.VITE_POSTHOG_HOST });
+    posthog.identify({
+        distinctId: userId,
+        properties
+    });
+    logger.debug(`identified posthoguser ${userId} with properties ${JSON.stringify(properties)}`);
+    await posthog.shutdown();
+}
+