+++ apps/src/routes/admin/jadwal/+page.server.ts
@@ -5,78 +5,87 @@ import { repo } from '$src/lib/server/db';
 import { redirect } from '@sveltejs/kit';
 import { logger } from '$src/lib/utils/logger';
 import { getWeekNumber, formatDate } from '$src/lib/utils/dateUtils';
-import { captureEvent } from '$src/lib/utils/analytic';
+import { handlePageLoad } from '$src/lib/server/pageHandler';
+import type { Event } from '$core/entities/Event';
 
 /**
  * Page server load function for the jadwal (schedule) page.
  * @type {import('./$types').PageServerLoad}
  */
-export const load: PageServerLoad = async (events) => {
+export const load: PageServerLoad = async (event) => {
 	// Check if the user is authenticated
-	const session = await events.locals.auth();
+	const { session } = await handlePageLoad(event, 'jadwal');
 
 	if (!session) {
 		throw redirect(302, '/signin');
 	}
 
 	const churchId = session.user?.cid ?? '';
+	let masses: any[] = [];
+	let eventsDetail: any[] = [];  // TODO: not so type safety, Event has no ushersCounts
 
 	// Initialize services
 	const churchService = new ChurchService(churchId);
 	const eventService = new EventService(churchId);
 
 	// Fetch masses and events concurrently
-	const [masses, massEvents] = await Promise.all([
-		churchService.getMasses(),
-		churchService.getEvents()
-	]);
+	try {
+		const [fetchedMasses, massEvents] = await Promise.all([
+			churchService.getMasses(),
+			churchService.getEvents(10)
+		]);
 
-	// Process each event to include detailed data
-	const eventsDetail = await Promise.all(
-		massEvents.map(async (event) => {
-			// Fetch ushers for the event
-			const ushers = await eventService.getEventUshers(event.id);
+		masses = fetchedMasses;
 
-			// Get mass details and positions
-			const massId = await repo.getEventById(event.id);
-			const requiredPositions = await churchService.getPositionsByMass(massId.mass);
-			const totalUshers = requiredPositions?.length ?? 0;
+		// Process each event to include detailed data
+		eventsDetail = await Promise.all(
+			massEvents.map(async (event: Event) => {
+				// Fetch ushers for the event
+				const ushers = await eventService.getEventUshers(event.id);
 
-			// Calculate usher statistics
-			const confirmedUshers = ushers?.length ?? 0;
-			const totalPpg = ushers.filter((usher) => usher.isPpg).length ?? 0;
-			const totalKolekte = ushers.filter((usher) => usher.isKolekte).length ?? 0;
+				// Get mass details and positions
+				const massId = await repo.getEventById(event.id);
+				const requiredPositions = await churchService.getPositionsByMass(massId.mass);
+				const totalUshers = requiredPositions?.length ?? 0;
 
-			// Keep progress between 0 and 100
-			const progress = Math.min((confirmedUshers / totalUshers) * 100, 100);
+				// Calculate usher statistics
+				const confirmedUshers = ushers?.length ?? 0;
+				const totalPpg = ushers.filter((usher) => usher.isPpg).length;
+				const totalKolekte = ushers.filter((usher) => usher.isKolekte).length;
 
-			// Return event with additional usher count information
-			return {
-				...event,
-				usherCounts: {
-					progress: progress,
-					totalUshers: totalUshers,
-					confirmedUshers,
-					totalPpg,
-					totalKolekte
-				}
-			};
-		})
-	);
+				// Keep progress between 0 and 100
+				const progress = Math.min((confirmedUshers / totalUshers) * 100, 100);
+
+				// Return event with additional usher count information
+				return {
+					...event,
+					usherCounts: {
+						progress: progress,
+						totalUshers: totalUshers,
+						confirmedUshers,
+						totalPpg,
+						totalKolekte
+					}
+				};
+			})
+		);
+	} catch (error) {
+		logger.error('Error fetching masses and events:', error);
+	}
 
 	// Filter events into this week and past events based on week number
 	const currentWeek = getWeekNumber(new Date().toISOString());
-	const thisWeekEvents = eventsDetail.filter(event => event.weekNumber === currentWeek);
-	const pastEvents = eventsDetail.filter(event => (event.weekNumber ?? 0) < currentWeek);
+	const thisWeekEvents = eventsDetail.filter((event) => event.weekNumber === currentWeek);
+	const pastEvents = eventsDetail.filter((event) => (event.weekNumber ?? 0) < currentWeek);
 
 	const activityItems = pastEvents.map(event => ({
-		title: `<a href="/admin/jadwal/${event.id}" class="font-semibold text-primary-600 dark:text-primary-500 hover:underline">${event.mass}</a>`,
+		link: `<a href="/admin/jadwal/${event.id}" class="font-semibold text-primary-600 dark:text-primary-500 hover:underline">${event.mass}</a>`,
 		date: formatDate(event.date),
-		src: '',
-		alt: ''
+		church: event.church,
+		churchCode: event.churchCode,
+		mass: event.mass,
+		progress: Math.round(event.usherCounts.progress)
 	}));
-
-	await captureEvent(events, 'jadwal_page_view');
 
 	// Return masses and processed events
 	return {