+++ apps/src/routes/admin/misa/+page.server.ts
@@ -1,14 +1,15 @@
 import type { Mass } from '$core/entities/Schedule';
 import { ChurchService } from '$core/service/ChurchService';
-import { captureEvent } from '$src/lib/utils/analytic';
+import { handlePageLoad } from '$src/lib/server/pageHandler';
+import { captureEventServer } from '$src/lib/utils/analytic';
 import { logger } from '$src/lib/utils/logger';
 import type { PageServerLoad } from './$types';
 import { redirect } from '@sveltejs/kit';
 
-export const load: PageServerLoad = async (events) => {
+export const load: PageServerLoad = async (event) => {
     logger.info('Loading page admin/misa');
 
-    const session = await events.locals.auth();
+    const { session } = await handlePageLoad(event, 'misa');
     if (!session) {
         logger.error('User not authenticated');
         throw redirect(302, '/signin');
@@ -19,7 +20,7 @@ export const load: PageServerLoad = async (events) => {
 
     const masses = await churchService.getMasses();
 
-    await captureEvent(events, 'misa_page_view');
+    // await captureEvent(events, 'misa_page_view');
 
     return {
         masses