+++ apps/CHANGELOG.md
@@ -1,3 +1,24 @@
+## 0.0.1 (2025-03-21)
+
+
+### Bug Fixes
+
+* assignment without positions ([4665a68](https://github.com/sutanto62/lilium/commit/4665a682a36469f2f7148d3d595d495eb4eabdda))
+* auth provider error ([01c315d](https://github.com/sutanto62/lilium/commit/01c315dd9899f8f851b0967a552ea8d430647158))
+* error svelte-tags-input ([fceab41](https://github.com/sutanto62/lilium/commit/fceab41277c5b060dd539ef94bd96137beb92fdc))
+* event created without position ([#7](https://github.com/sutanto62/lilium/issues/7)) ([#8](https://github.com/sutanto62/lilium/issues/8)) ([52de11b](https://github.com/sutanto62/lilium/commit/52de11bb92b41d46a45c581c8682e82e690d54e7))
+* pic layout ([c20029f](https://github.com/sutanto62/lilium/commit/c20029f3ecce2f0e6feb661b111ea282b84d52af))
+* round robin with PPG positions first ([3be6316](https://github.com/sutanto62/lilium/commit/3be6316867bb22b738be3b942422ba23895927e1))
+* sso button ([#9](https://github.com/sutanto62/lilium/issues/9)) ([db9a740](https://github.com/sutanto62/lilium/commit/db9a74010f73fb142399a6aeb90ab342f1d32cd4))
+
+
+### Features
+
+* download usher list ([e7fc977](https://github.com/sutanto62/lilium/commit/e7fc977731b92f766a08cd63f0814c132464df6c))
+* event and usher management ([284d99a](https://github.com/sutanto62/lilium/commit/284d99a4aaec158144836e510da4086b10e58181))
+
+
+
 ## 0.0.1-alpha (2024-12-30)
 
 ### Bug Fixes